backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
pillow==10.3.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
tensorflow==2.15.0
deep-translator==1.11.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
joblib==1.4.2
